refactor: Separate Card.module.scss from Home.module.scss
