+++ components/card/Card.tsx
@@ -4,7 +4,7 @@ dayjs.extend(timezone);
 import timezone from 'dayjs/plugin/timezone';
 import utc from 'dayjs/plugin/utc';
 import Link from 'next/link';
-import styles from '@/styles/Home.module.scss';
+import styles from './Card.module.scss';
 import { Post } from '@/types/blog';
 
 export default function Card({ post }: { post: Post }) {
