+++ pages/index.tsx
@@ -5,9 +5,9 @@ import timezone from 'dayjs/plugin/timezone';
 import utc from 'dayjs/plugin/utc';
 import { GetStaticProps } from 'next';
 import Head from 'next/head';
-import Card from '@/components/Card';
 import Footer from '@/components/Footer';
 import Header from '@/components/Header';
+import Card from '@/components/card/Card';
 import { Pagination } from '@/components/pagination/Pagination';
 import { microcms } from '@/libs/microcms';
 import styles from '@/styles/Home.module.scss';
