+++ styles/Home.module.scss
@@ -1,49 +1,49 @@
 @import './variables';
 
 .container {
-  align-items: center;
   display: flex;
   flex-direction: column;
+  align-items: center;
   justify-content: center;
   min-height: 100vh;
   padding: 0 8px;
 }
 
 .main {
-  align-items: center;
   display: flex;
   flex: 1;
   flex-direction: column;
+  align-items: center;
   justify-content: center;
-  padding: 80px 0;
   width: 100%;
+  padding: 80px 0;
 }
 
 .footer {
-  align-items: center;
-  border-top: 1px solid $light-gray;
   display: flex;
-  height: 100px;
+  align-items: center;
   justify-content: center;
   width: 100%;
+  height: 100px;
+  border-top: 1px solid $light-gray;
 
   span {
     display: flex;
   }
 
   a {
-    align-items: center;
-    color: inherit;
     display: flex;
     flex-grow: 1;
+    align-items: center;
     justify-content: center;
+    color: inherit;
     text-decoration: none;
   }
 }
 
 .title {
-  font-size: 4rem;
   margin: 0;
+  font-size: 4rem;
   text-align: center;
 
   a {
@@ -64,41 +64,11 @@
 }
 
 .grid {
-  align-items: center;
   display: flex;
   flex-wrap: wrap;
+  align-items: center;
   justify-content: center;
   width: 80%;
-}
-
-.card {
-  border: 1px solid $light-gray;
-  border-radius: 10px;
-  color: inherit;
-  margin: 16px 0;
-  padding: 24px;
-  text-align: left;
-  text-decoration: none;
-  transition: color 0.15s ease, border-color 0.15s ease;
-  width: 100%;
-
-  h2 {
-    font-size: 1.5rem;
-    margin: 0;
-    margin-bottom: 16px;
-  }
-
-  p {
-    font-size: 1rem;
-    margin: 0;
-  }
-
-  &:hover,
-  &:focus,
-  &:active {
-    border-color: $primary-purple;
-    color: $primary-purple;
-  }
 }
 
 .twitterAccount {
